feat(docs): remove license mention from readme file
The license mention is redundant, since the LICENSE file is shown
in a separate tab in GitHub repository overview page, and stated
in repository's attributes as well.

On the other hand, most likely the inherited repositories will
change the license to CC-BY-x or other. Having linked file may
cause issues during license change.